Fix typo in "WepApi" by renaming to "WebApi" across tests
Renamed files, namespaces, and references in the test project to correct the typo in "WepApi" to "WebApi". This ensures consistency and accuracy throughout the test files and related configurations in the solution.
