+++ Postech.Fiap.Orders.sln
@@ -11,7 +11,7 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution 
 		docker-compose.yaml = docker-compose.yaml
 	EndProjectSection
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Postech.Fiap.Orders.WepApi.UnitTests", "test\Postech.Fiap.Orders.WepApi.UnitTests\Postech.Fiap.Orders.WepApi.UnitTests.csproj", "{06764304-80BE-4641-80D5-9AA76BCFC2D4}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Postech.Fiap.Orders.WebApi.UnitTests", "test\Postech.Fiap.Orders.WebApi.UnitTests\Postech.Fiap.Orders.WebApi.UnitTests.csproj", "{06764304-80BE-4641-80D5-9AA76BCFC2D4}"
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "CI/CD", "CI/CD", "{FBD2C357-44B6-4CFB-A991-5117B78529FD}"
 EndProject
