--- test/Postech.Fiap.Orders.WepApi.UnitTests/Features/Orders/Entities/OrderIdTests.cs
@@ -1,32 +1,0 @@
-using Postech.Fiap.Orders.WebApi.Features.Orders.Entities;
-
-namespace Postech.Fiap.Orders.WepApi.UnitTests.Features.Orders.Entities;
-
-public class OrderIdTests
-{
-    [Fact]
-    public void OrderId_Should_Store_Guid_Value_Correctly()
-    {
-        // Arrange
-        var guid = Guid.NewGuid();
-
-        // Act
-        var orderId = new OrderId(guid);
-
-        // Assert
-        orderId.Value.Should().Be(guid);
-    }
-
-    [Fact]
-    public void New_Should_Generate_Unique_Guid()
-    {
-        // Act
-        var orderId1 = OrderId.New();
-        var orderId2 = OrderId.New();
-
-        // Assert
-        orderId1.Should().NotBe(orderId2);
-        orderId1.Value.Should().NotBe(Guid.Empty);
-        orderId2.Value.Should().NotBe(Guid.Empty);
-    }
-}
